BlogProject/requirements_local.txt
asgiref==3.8.1
attrs==24.1.0
autobahn==24.4.2
Automat==22.10.0
certifi==2024.8.30
cffi==1.16.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-oauth-toolkit==2.4.0
django-redis==5.4.0
djangorestframework==3.15.2
dnspython==2.6.1
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jwcrypto==1.5.6
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==24.2.1
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==6.4.post2
